refactor(asset-collection): IErrorArtifactRepository を application 層に移動する
infrastructure 層にあった IErrorArtifactRepository / ErrorArtifactUploadError を
application 層に移動し、クリーンアーキテクチャの依存方向を正す。

docs: CONTEXT.md、ADR-0001、CONTRIBUTING.md のスクレイピング層責務を追加
