+++ lambda/asset-collection/src/application/__init__.py
@@ -1,10 +1,13 @@
 from .asset_fetcher_interface import AssetFetchFailed, IAssetFetcher
 from .collect_asset_interface import ICollectDailyAssetUseCase
 from .collect_asset_usecase import CollectAssetDailyUseCase
+from .error_artifact_repository_interface import ErrorArtifactUploadError, IErrorArtifactRepository
 
 __all__ = [
     "AssetFetchFailed",
-    "ICollectDailyAssetUseCase",
-    "IAssetFetcher",
     "CollectAssetDailyUseCase",
+    "ErrorArtifactUploadError",
+    "IAssetFetcher",
+    "ICollectDailyAssetUseCase",
+    "IErrorArtifactRepository",
 ]
