+++ lambda/asset-collection/src/application/error_artifact_repository_interface.py
